pyproject.toml
[tool.poetry]
name = "boot-habito-whatsapp"
version = "0.1.0"
description = ""
authors = ["Paulo Victor <paulovictorsantosvc@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
twilio = "^9.4.4"
flask = "^3.1.0"
pandas = "^2.2.3"
schedule = "^1.2.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
